fix: get_panics_when_out_of_bounds test, release build
This test was failing with `cargo test --release`.

The test calls get_leafs_async() which panics in a debug_assert!(false).  The
test then verifies the panic error message.

An equivalent but slightly different panic occurs for the release build,
which doesn't match the expected panic string.

For now, I have simply decorated the test with #[cfg(debug_assertions)] so it is
only compiled for debug builds.