build: add tokio-console feature flag.
closes #379

The tokio-console feature requires dep console-subscriber, which pulls
in over 100+ additional dependencies, including things like hyper and
axum.

This commit adds a tokio-console feature flag and makes the
console-subscriber dep optional.

Here we can see the reduction in deps:

$ cargo tree -e no-dev --features tokio-console | wc -l
874

$ cargo tree -e no-dev | wc -l
735

To be clear: this means that tokio-console support is no longer built
by default.  It requires building with --features tokio-console.