fix(`NativeCurrencyAmount`): Propagate rounding carry
When cast to a string with a small (<34) number of decimal places
after the comma, native currency amounts lose precision and rounding
occurs. Rounding down has no effect; rounding up induces a carry
chain. Previously the rounding was limited to the last digit. If
the last digit is "9" and rounding occurs, the last digit becomes
"10" which is wrong because it is not one digit but two.

The fix is to wrap around the base of expansion and carry the
increment-by-one to the next digit in the direction of greater
significance. So "9" becomes "0" and the digit to the left is
incremented by one -- unless it is a "9" already in which case it,
too, wraps around and the carry chain keeps propagating. If the
carry chain reaches the integral part, it increments that integer
and terminates there because the integral part is not bound to
0..=9.

Fixes #383.