fix(wallet): Don't create transaction using UTXOs present in mempool
When two transactions are created in short succession, the 2nd
transaction would, prior to this commit, select the same input UTXOs as
the 1st transaction. This prevents both transactions from a) being
mined, and b) living in mempool simultaneously. With this commit, the
input-UTXOs are checked to see if they match a transaction in the
mempool. If they do, then the UTXO is skipped, and the next one is
considered.

This closes #372.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>