refactor(SingleProof): Audit PC witness at end of program
This branch of SingleProof has a lot of reliance of memory integrity of
the witness -- that the witness preloaded into memory does not get
modified during program execution. This extra audit at the end helps
make a successful modification of the witness quite unlikely, we
believe.

If you disagree, why don't you see if can design a test case that
returns the new error code.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>