feat(`WalletState`): Automatically detect own guesser UTXOs
Previously, detection of guesser fee UTXOs from own guesses required
updating wallet state with a list of `ExpectedUtxo`. This led to the
situation where guesser fee UTXOs were not recognized by a different
machine that uses the same wallet secret.

As guesser fee UTXOs are locked by a `RawHashLock` whose unlock key is
a deterministic function of the wallet secret and the previous block's
digest, the wallet might as well reproduce this unlock key and match
new blocks' guesser fee UTXOs against this value. This commit achieves
just that.

The mechanics for transmitting the guesser preimage in case of a found
block are now superfluous. These will be removed in a future commit.