refactor: Avoid passing guesser preimages
When a nonce is found, the guesser sends a message upstream to
notify its elders that a block has been found. This message used to
contain the guesser preimage.

In practice, the guesser preimage was always set to a value
deterministically derived from the wallet secret and the previous
block digest. The previous commit modifies the wallet update function
to check for guesser fee UTXOs that can be unlocked by this
deterministic preimage (in addition to any expected UTXOs). As a
consequence, the mechanism for transmitting the guesser preimage was
made entirely redundant.

This commit eliminates it.