chore: add php script to calculate block proposal durations
The script will parse an input neptune-core logfile to
find new block events and the first subsequent block-proposal
event.   It then calculates the time difference and reports.

It is imperfect in that the block proposal log message does not
report a height, so they could be mis-matched, but that should be
rare.