feat: Refuse connections from recent disconnects
In order to reduce the potential connection flood, rapid re-connection
attempts after a graceful disconnect are ignored.

If node A decides to (gracefully) terminate the connection to some node
B, then node A blocks any connection attempts from node B for some
time. The cool-down time can be configured through the command-line
interface's argument “reconnect_cooldown”. The argument is given in
seconds. The default is 30 seconds.