feat: add peer column sorting
addresses #455

implements sorting of all columns in the peers screen.

also refactors a bit to enable cli args to be passed into peer screen
or other screens in the future if necessary.

Each column can be sorted invidually.  Also the sort order can be
chosen, either ascending or ascending.

Two methods are available for setting the sort order:

1. 2 cli flags are added, to set initial sort column and order.
2. hot keys are enabled for controlling the sorting interactively.

The hot keys are printed at top of screen, to make usage obvious.