feat: improve interactive peers sorting
addresses #455

previously one would sort a column using the current sort-order by pressing its
hotkey, eg 's' for the 'standing' column.
Then to change the sort order, press 'a' or 'd'.

That was a little tedious.

Now the hot-keys are easier to use.  Just press 's' to sort the 'standing' column
and then press 's' again to reverse the sort order.  any number of times.