chore: do not log slow scope across peer.send()
addresses #457

The log_slow_scope!() macro was being invoked for all peer message
variants. In some cases the measured scope was calling
peer.send().await which can take an indefinite amount of time to
transmit data to the peer.  The results of such measurements are not
fit for purpose, with lots of false positives.

This commit avoids measuring across peer.send().await.  In some
cases the rest of the message processing is measured.  In other cases
where the code is too complex with send() sprinkled throughout, the
measuring is removed entirely.