refactor: Reduce ways of setting new tip to one
Delete the method `set_new_self_mined_tip`/`set_new_self_composed_tip`
(name was in flux). Now there is only one method on `GlobalState` to set
a new tip. This means that it is the caller's responsibility to tell the
state to expect any transactions before the state-update is applied.

This responsibility was already managed by the non-testing (mainline)
code; the mining logic already handle self-produced block compositions
in a way consistent with this, since the main_loop already added
composer UTXOs to its list of expected UTXOs prior to this commit.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>