bench: Add PoW benchmark showing hash rate
Add a benchmark that should give a good estimate of your machine's
maximum hash rate. Assuming 400 seconds of guessing time, you can
multiply the reported number with 400 to find the difficulty that *your*
machine can solve. Divide this number by the reported difficulty to find
your probability of finding the PoW solution to the next block.

On my machine, I can make 10^6 guesses in ~65ms. This translates to a
guess-rate of 10^6/0.065 ~= 15.4 * 10^6 guesses per second, or 15.4
Mega-guesses per second.