refactor: Streamline transaction creation interface
Add `TxCreationConfig` struct to configure options when creating
transactions. Add `TxCreationArtifacts` struct to manage stuff
produced when creating a transaction. Use these objects to simplify
interface of `create_trasnaction`.

The config allows the user to specify a criterion by which to select
UTXOs. The artifacts object can contain (if the config says to
populate) a hash set of selected UTXOs.

Also:
 - Add test for size statistics on increasingly large blocks. This
   test requires the criterion and UTXO selection, so as to avoid
   selecting the same UTXO again for the next transaction.
 - Drop `create_transaction_with_prover_capability` as this
   indirection is now a configurable option.
 - Add test showing sizes of random transaction kernels.