refactor: Streamline transaction creation interface
Add `TxCreationConfig` struct to configure options when creating
transactions. Add `TxCreationArtifacts` struct to manage stuff
produced when creating a transaction. Use these objects to simplify
interface of `create_trasnaction`.

The config allows the user to specify a criterion by which to select
UTXOs. The artifacts object can contain (if the config says to
populate) a hash set of selected UTXOs.

Also:
 - Add test for size statistics on increasingly large blocks. This
   test requires the criterion and UTXO selection, so as to avoid
   selecting the same UTXO again for the next transaction.
 - Drop `create_transaction_with_prover_capability` as this
   indirection is now a configurable option.
 - Add test showing sizes of random transaction kernels.

chore: Fix rebase artifacts

fix: Wallet state maintenance test

Use `SingleProof` as previously.

Also: correct spelling mistake.

refactor: Wrap `TritonVmJobQueue` in smart pointer

The job queue was already being passed everywhere by reference. This
commit replaces the `&` reference with an `Arc<..>`.

Having an `Arc` reference is more conducive to storing the reference,
for instance in configuration objects such as `TxCreationConfig`.
Previously, that struct required complex lifetime annotations.
Moreover, the reference there was already (and still) wrapped in an
`Option` but because of the difficulty of storing `&`-reference, there
was no way of defaulting to a sensible default in case of `None`.

As a consequence, this change obviates the need to supply the `dummy`
job queue in all tests that use that. Less lines yay!

style: Follow linter recommendations

 - Use `.copied()` where possible.
 - Drop needless `continue`s.

style: Follow reviewer recommendations

Co-authored-by: Ferdinand Sauer <ferdinand@neptune.cash>

test: Fix

Fixes test `allow_consumption_of_genesis_output_test`.

Use the same configuration wrt on/off chain notifications as were
available previously. It is not clear that this "fixes" the test
because it is not clear there was an error to begin with. However,
with the wrong configuration (which is the thing that this commit
fixes), the test runner will start proving and on weak machines
(like CI) fail to complete that task.

test: Drop informational tests

Tests `block_size_limit` and `pseudorandom_transaction_kernel_sizes` did
not actually test anything, but instead reported size information. The
latter test has been dropped. The former test has been refactored into
one that exhibits the strategy for avoiding reuse of the same UTXOs when
constructing a big transaction from many small ones.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>

refactor: Allow burning change

Co-authored-by: Alan Szepieniec <alan@neptune.cash>

style: Make compiler happy

Co-authored-by: Alan Szepieniec <alan@neptune.cash>

refactor: Use mempool to avoid double-spends

Drop the field `selection` from `TxCreationArtifacts` as the task of
keeping track of unconfirmed spent UTXOs can be handled by the mempool

style: Drop `Box`

Specifically, `&Box<..>` was changed into `&..` because of a linter
recommendation.

refactor: Drop configurable filter to select UTXOs by

Filter was unused. Change simplifies structures and function
interfaces.

That said, the deletions from this commit may be a useful pattern for
future matching, for instance if we want to pass a closure with to
sort the UTXOs before selecting the first enough-many.

docs: Fix links or drop them

style: Integrate reviewer feedback

Co-authored-by: Ferdinand Sauer <ferdinand@neptune.cash>

style: Fix linter error