fix(primitive_witness): Actually check TVM result in validation
Prior to this commit, the actual Triton VM result was not checked.
The type being checked was `Result<Result<T>>` where only the outer
`Result` was checked for being OK. But the outer `Result` came from the
spawning of a tokio task, not from Triton VM, as the code assumed.

This clusterfuck lead to false positives in the validity check of
primitive witnesses. Since a PW-backed transaction cannot be mined or
merged though, this error has had no effect on the soundness of the
blockchain. But it has negatively impacted the quality of some tests.

Oh yeah, btw., the state setup for executing type scripts in Triton VM
in the context of PW-verification was also completely wrong, as the
Triton VM execution result was never actually checked -- so that also
went unnoticed.

This shows the importance of negative tests. Always, always, always
write some negative tests of your verfication functions.

To that effect: Add negative tests for catching a similar regression,
for both lock script validity and type script validity.