feat(rpc_server): Add endpoint for external guesser, with external key
Add an endpoint allowing the exporting of a PoW puzzle with a guesser
preimage *not* known by the node software.

This allows for "cold" guessing where the node software (which might be
running on other people's computers AKA the "cloud") never learns secret
information required to unlock a guesser reward.

Since this can be used for providing other people the service of block
proposals, we cap the number of cached block proposals to 10'000 as this
would eat up 10-20GB RAM.