refactor: Wrap `TritonVmJobQueue` in smart pointer
The job queue was already being passed everywhere by reference. This
commit replaces the `&` reference with an `Arc<..>`.

Having an `Arc` reference is more conducive to storing the reference,
for instance in configuration objects such as `TxCreationConfig`.
Previously, that struct required complex lifetime annotations.
Moreover, the reference there was already (and still) wrapped in an
`Option` but because of the difficulty of storing `&`-reference, there
was no way of defaulting to a sensible default in case of `None`.

As a consequence, this change obviates the need to supply the `dummy`
job queue in all tests that use that. Less lines yay!