test: Fix
Fixes test `allow_consumption_of_genesis_output_test`.

Use the same configuration wrt on/off chain notifications as were
available previously. It is not clear that this "fixes" the test
because it is not clear there was an error to begin with. However,
with the wrong configuration (which is the thing that this commit
fixes), the test runner will start proving and on weak machines
(like CI) fail to complete that task.