fork!: Increase block size
We discovered that the number of inputs allowed into a block had a much
lower practical limit than we intended. The problem is that the `inputs`
field of the `TransactionKernel` holds up to 45 authentication paths of
the sliding-window Bloom filter MMR. If this MMR has a height of 20,
which will be reached when the total number of outputs is ~8M, these 45
authentication paths take up 4.500 `BFieldElement`s. The block proof
being ~125.000 `BFieldElement`s and the old block size limit 250.000,
this only leaves room for 125.000 / 4.500 ~= 28 inputs! And that's
assuming that the mutator set accumulator takes up no space, and that
there are no outputs.

With this increase, the practical limit on the number of inputs per
block rises to ~195, which is still quite low but should be enough for
"some" time.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>