feat: Allow restriction of number of inputs per tx
Allow the user to restrict the number of inputs per transaction to avoid
having to deal with transactions that will likely never be mined, as
they don't fit into a block.

The current default limit is set quite low, to 28 inputs per
transaction due to the big size of transaction inputs. This limit is
intended to rise significantly in the future, but for now it should
serve to prevent composers and transaction initiators from making
transactions that cannot be mined.

Both the wallet state and mempool state gets passed this value such that
internally can enforce the correct policy.