fix(peer_loop): fix race condition on new connection
Two tokio tasks can run simultaneously that establish a connection
incoming/outgoing to the same peer. If a write lock is not held over
some final checks before a connection is established then two
connections to the same peer might be made.

This reintroduces the holding of a lock while this final check is being
made. The lock was removed in commit fa89d0414.