fix(mempool): Ignore conflictors with deprecated mutator sets
The policy for rejecting an insertion into the mempool is modified.
Existing transactions that are in conflict with the new transaction
and which have a higher fee density are kicked out anyway (contrary
to before) if the new transaction has identical inputs and outputs.
That identity is taken as an indicator that the new transaction is
the result of updating an old transaction's mutator set, in which case
we do not care about the deprecated transaction any more.

This policy change was necessary for fixing the upgrader fee recovery
tests, which now pass.