refactor(`TritonVmJobQueue`): Use singleton pattern
The purpose of the job queue is to prevent different jobs from
competing for the same compute resources at the same time. As such,
having multiple job queues defeats the purpose of having one. In other
words, there should only ever be one job queue.

This commit refactors the interface of `JobQueue` and
`TritonVmJobQueue` to enforce that only one instance of the latter
type is ever constructed. Specifically, the constructor `start` is
downgraded in visibility from public to private. To get the Triton VM
proof queue object, the caller calls `global_triton_vm_job_queue()`,
which is a public function that internally uses a static `OnceLock` to
ensure that no more than one instance is ever created.

Also, this commit removes the reference to the triton vm job queue
from all function interfaces.