style: Report guesser fee of rejected blocks
When rejecting a block proposal notification, an info message is
logged showing the reason why it was rejected. This commit changes
that info message to include the rejected proposal's guesser fee.

This change answers a request made by a composer who would like to
know what fees other competing composers are offering, in order to
adapt their own price.