refactor(`mine_loop`): Factor out extractor for `ExpectedUtxo`s
At some point in the mine loop, the composer has to extract a list of
`ExpectedUtxo`s from a `TxOutputList`, but only if a) the fee
notification policy is offchain, and b) it is not cold-composing. This
extraction step was previously inline and duplicated. It is now factored
out into a function living on `ComposerParameters`, which contains the
conditionals and calls another function on `TxOutputList`, which
performs the conversion.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>