refactor(`mempool`): Catch self-conflicts smarter
Upon inserting a transaction into the mempool, all conflicting
transactions are kicked out if:
 - they have inferior proof quality; or
 - they have lesser fee density.

This commit expands the definition of proof quality to include
updates of the mutator set data. The end result is that transactions
with updated mutator set data are always allowed to replace their
progenitors. (Note that this behavior was achieved with an explicit test
in a prior commit.)

Co-authored-by: Alan Szepieniec <alan@neptune.cash>