privacy(mine_loop): Never add zero-valued composer outputs
If a block is empty, i.e. contains no inputs and only the coinbase
rewards, *and* the composer chooses to reward the entire block subsidy to
the guesser, there is no need for the composer to add any outputs to the
coinbase transaction.

Prior to this commit, the composer would in that scenario add two ouputs
to the coinbase transaciton. Many such blocks with two zero-valued outputs
exist on main net today, so this is not a hypothetical scenario.

Avoiding these zero-valued outputs has two benefits:
 - It leaves more room for other transactions to be merged into the
   block.
 - It avoids zero-valued leafs in the AOCL whose presence would otherwise
   reduce the anonymity set as zero valued leafs are less likely to be
   spent than any other leafs and can, in the case of an empty block, be
   identified in the AOCL.