fix(peer_loop): Don't hold read lock responding with transction
Prior to this commit, a read-lock was held while a message was sent to a
peer. That is totally disallowed since the connection can timeout in
which case this read lock will be held for up to two minutes.

Any and all locks must be released before sending anything to a peer!

Cf. #534.