test(proof_upgrader): Test merge-output being double-spender
Add a test of the scenario where one of the two inputs to a
merge-proof-upgrade is mined while the merge proof production is
happening. In this scenario, the client is not allowed to share the
transaction with peers since it includes inputs that have been spent and
it is thus a double-spending transction. The correct behavior is to
discard the client's work.

The test verifies both the correct behavior of the main-to-peer-loop
channel and that of the mempool.

fixup test-logic