refactor: pass network to proof verify function
replaces the Network singleton previously used for this purpose.

The singleton approach applies the network process-wide.  That means
that it is unsuitable for use in unit-tests which set different
networks.  it is also problematic for integration tests.  And it would
likely be problematic for any third parties that use neptune_cash as a library.

The prior code worked around this limitation by only setting the
singleton in two places:

1. in main()
2. in GenesisNode::start_node() used by integration tests.

This worked for now, so long as all integration tests use regtest.  But
there remains potential for mis-match between the network singleton and
the cli_args::Args network. It is not a sound solution.

Alternative Approaches:

1. pass "allow_mock_proof" param to verify instead of network.  This
   works the same.

2. perform the Network::use_mock_proof() check in the peer loop when accepting
   any Proof from a peer.  The verify() function would not receive a network
   param and would simply trust that mock proofs are always allowed.

This commit opts to pass network to verify() because:

* network is also passed to prove() and prove_concensus_program() via
  proof_job_options parameter.  So it is (more) symmetric to also pass it
  verify().
* performing the check within verify() is more robust in terms of
  security.  The verify() fn need not rely on peer_loop checking all
  incoming proofs, or other possible attack vectors such as deserialization,
  on-disk shenanigans, etc.