feat: implement mockable proofs for regtest mode
Purpose: this commit enables mock proofs (valid or invalid) to be sent between
neptune-core peers, which makes regtest mode usable across a network.

`MockableProof` is a newtype for `triton_vm::proof::Proof` that can
represent three logical variants: NotMock, ValidMock, InvalidMock.

The NotMock variant is identical to a triton_vm::proof::Proof including
when serialized and deserialized.  So the wire format does not change.

ValidMock represents a mock-proof that will validate.
InvalidMock represents a mock-proof that will fail validation.

The ValidMock and InvalidMock variants encode the mock-proof behavior
inside the Vec<BFieldElment> of the triton-vm Proof.  When a
MockableProof is deserialized this information is retained and is used
to determine if the proof is either of the mock variants.

Note also that if a ValidMock were ever to be deserialized as a standard
triton_vm::proof::Proof it would fail validation.  Which is good.

Changes:

1. Add MockableProof
2. type alias Proof to MockableProof.
3. Make Network a singleton.
4. Verify proofs as mock only if network permits it

Squashed commits:

* args singleton -> network singleton.
* cli_args singleton --> network singleton
* manually impl BFieldCodec for MockableProof.  all tests pass
* fix CLAIMS_CACHE comment
* encode behavior inside existing proof elements. avoids changing wire format
  for non-mock proofs
* cleanup, comments
* prove regtest mock proofs in prover_job
* impl BFieldCodec manually again. otherwise the ProverJob::check_if_allowed()
  fails for some reason
* revert to gen mock proof in builder, not in ProverJob.  for now.
* add mine-blocks-to-wallet neptune-cli command
* manually impl TasmObject again.  seems to be necessary
* set Network singleton from main(), so triton-vm verify() works in regtest mode