fix: make new-self-guessed-block handling atomic
a previous commit caused a regression by placing the is_canonical
check and set_new_tip() call in separate lock acquisitions, allowing
a race condition where a non-canonical block could become the tip.

This commit:
1. makes them atomic again.
2. holds the write-lock for less time.
3. broadcasts to peers after setting new tip, instead of before.
4. documents everything clearly.

(3) may not be strictly necessary as the send() call is not async, but
is precautionary.