refactor: Simplify TX interface
style: Simplify tx-related interfaces

Define two helper functions, one for deriving a transaction kernel from
`TransactionDetails`, and on for simpifying merge-proof generation.

Co-authored-by: Thorkil Schmidiger <thor@neptune.cash>
Co-authored-by: Alan Szepieniec <alan@neptune.cash>

refactor(`NativeCurrency`): Factor out amount logic

Consensus program `NativeCurrency` is refactored so that `Snippets`
of logic pertaining to amounts now live in their own files. These
snippets can now be used elsewhere, for instance by user-defined
type scripts that also check amounts (such as tokens).

The challenge was to ensure that the type script hash does not
change under this refactor. This is ensured by bypassing the
standard `Library::import` method of importing dependencies, which
cannot be used because it prefers its own order of dependencies.
While the order in which the dependencies are presented in the
program does not matter for the program's correctness, it does
change the program hash.

Third-party type script developers can and should use
`Library::import` because they are not bound by a type script hash
that is not allowed to change.

Also, the relevant snippets take an argument specifying how to
produce the type script hash. Typically this hash is the program
hash of whichever program is currently running (*e.g.*, the
`NativeCurrency` type script). However, in some cases users might
want to read foreign amount-like type script states and to enable
this use case the caller can also supply a hardcoded hash. For
example, a "vault" type script must reject spending attempts that
spend the entire treasury to an unrecognized address in one go but
allow small payments. To do this it must read the type script
states of `NativeCurrency` coins and do amount arithmetic on them
even though its own hash is not the `NativeCurrency` hash.

feat: Add snippet for total amount from UTXO list

This commit implements a snippet that reuses the snippets factored
out in the previous commit (*i.e.*, factored out from
`NativeCurrency`) into a standalone `Snippet` for computing the
total amount and total timelocked amount from a list of UTXOs
stored in memory.

Also, the test serves as extra validation (if we needed any) that
tasm code is correct.

refactor(`GetTotalAndTimeLockedAmounts`): Collapse nested `if`-statements