refactor: split proof builder and use in codebase
1. ProverJobSettings

adds fields tx_proving_capability and proof_type.  In conjunction with the
already-existing network field, this enables the proving job to validate if the
proof should proceed.

2. TransactionProofBuilder

fix: build mock-proofs of target proof-type

previously if the network is regtest, TransactionProofBuilder would always build
a mock SingleProof.  The corrected behavior is to build a mock proof of the
requested proof-type. (in particular ProofCollection).

other improvements to TransactionProofBuilder:
* build proof from single_proof_witness
* build proof from proof_collection
* build proof from claim-and-nondeterminism
* support both valid and invalid mock proofs
* logic is cleaner

3. ProofBuilder (new builder)

This is a builder for `Proof` directly from program, claim, and non-determinism.
It is aware of mock-proofs and can generate both valid and invalid mock proofs.

4. Codebase usage of Builders.

The proof-builders are used in the codebase replacing calls to produce(),
prove() and prove_concensus_program().  For more consistency.

5. Public API cleanups and improvements.

* made a few types public so they can be used properly with the
builders, eg ProverJobSettings
* added impl From<Args> for TritonVmProofJobOptions which makes it easy to do
builder.proof_job_options(cli.into())

squashed commits:

* feat: proof builder now supports valid_mock() option
* refactor TransactionProofBuilder and use it in upgrade_job
* hack: builder generates mock SingleProof any time mock proof requested
* replace an occurrence of SingleProof::produce() with proof-builder in
  mine-loop.  and warn if ::produce() is called in regtest mode
* support building single-proof from SingleProofWitness in
  TransactionProofBuilder, and use it in Transaction::merge_with(). should help
  with automated (real) mining in regtest mode
* add proof_builder, single_proof_builder, and use them in the codebase
* chore: support cli flags in script
* fix send_rate_limit with respect to tx_proving_capability