fix: integration test alice_sends_to_random_key logic
The test was previously relying on a mis-behavior (hack) in
TransactionProofBuilder whereby the builder always generates a SingleProof if
network is RegTest rather than producing a mock-proof of the requested
proof-type.

The problem was that the test-case was not waiting for the sent-tx to be
upgraded to single-proof in the mempool before generating the next block.

This commits adds GenesisNode logic so tests can easily wait for a tx to be
upgraded in the mempool.

The hack in TransactionProofBuilder is addressed in a future commit.