fix: disallow automated mining in regtest mode
Automated mining is not working with regtest mode yet.

But that's ok for now because regtest mode is intended for deterministic
testing with manually mined mock blocks.  Automated mining would be discouraged
in this mode even if it did work.

So until such time as we get it working, a temporary measure is to simply abort
execution if --compose or --guess is used with --network regtest.

An error is returned from initialize() and the program displays the error and
exits.