fix: enter scan mode only if wallet db not found
fixes #556

reverses the logic so that scan mode is entered when wallet db not found.

previously the wallet_is_new var was set to true if the wallet-db directory
exists. It must be false in that case.

This change also fixes an incorrect log message indicating the wallet db is not
found and wallet may have been imported.  This would rightly worry end-users.