fix: use correct table order in wallet db
addresses #552
addresses #553

somehow commit 2ace87ecb changed the order in which wallet db tables are loaded.

(which is ironic because it adds comments to NOT change the order)

The DB migration test passed because it also used the wrong order in the
simulated v0 database.

This commit fixes the order and also adds a test that uses a real v0 database
that was created by neptune-core v0.2.2 (tag) after sending a transaction.