fix: witness capable: upgrade proof in regtest mode
CI was failing for integration test alice_sends_to_bob because
the CI machine only has primitive-witness capability and the
proof-upgrader panics if capability is not at least
ProofCollection.

This changes the proof-upgrader logic so that if:

   capability is PrimitiveWitness && network.use_mock_proof()

then upgrade to SingleProof. otherwise panic as before.

An alternative would be to always create a SingleProof when sending for the
network.use_mock_proof() case, however then the entire proof upgrading path gets
bypassed, and I think it's better to exercise that code in regtest mode, so that
regtest mode stays as close as possible to the standard path.