test: fix race_condition_with_one_new_block
This test was hanging forever in CI.

There were two problems:

1. The CLI args were using the default auto-detected TxProvingCapability which
results in PrimitiveWitness capability on the CI machine which causes the proof
upgrader job to log an error and return (abort).

2. The test was passing a cloned channel sender to handle_upgrade() then calling
recv() on the receiver prior to dropping the original sender.  So after
handle_upgrade() aborts and returns the test waits for this message which will
never come because handle_upgrade() already dropped that sender.  But the test
is holding the original sender so it never gets dropped.  deadlock.

The fix for (1) is set the CLI args capability to match the capability in each
iteration of the loop [ProofCollection, SingleProof].

The fix for (2) is simply to *move* the sender into handle_upgrade() instead of
cloning it, so that it gets dropped when the fn ends.  THen the recv().await
immediately returns an error because the channel is closed.