test: use localhost to avoid windows firewall
Integration tests that involve one peer connecting to another were
failing on windows (only).

This was because windows firewall by default prevents (silently?)
application attempts to bind to public interfaces.  The default value
of cli_args::Args::listen_addr is "::" which means bind to all
interfaces, and thus the bind gets blocked, even for localhost.

The fix is to set listen_addr for tests to 127.0.0.1 before starting
the node.

with this change, the tests pass, and CI should succceed for windows
once again.