ci: Exclude test code from code coverage
Previously, test code was itself considered as code-under-test, skewing
the reported coverage percentage. Now, `#[cfg(test)]` modules are
explicitly excluded from code coverage.

The alternative tool `cargo tarpaulin` excludes tests in this way by
default. However, it seems to have trouble when multiple processes are
in play, as is the case with the `job_queue`.