fix: don't open DB before backup on windows
the backup db test was failing on windows because the backup function
opens the DB before/while copying. It was done that way purposely to
ensure exclusive access during the copy via leveldb's locking
mechanism.

But windows doesn't allow that and instead enforces exclusive access
via the filesystem.  So for windows we no longer open the DB.