chore: store db backups in migration_backups dir
purpose: stores db backups (prior to schema migration) in a
sub-directory "migration_backups" instead of in the same directory
as the in-use databases.

This directory is intended to be shared by all DBs that we use.

It is *not* intended to be shared by backups for other purposes.

Some refactoring of WalletConfiguration and DataDirectory was
required to do this cleanly, especially as this directory is not
specific to the wallet database.