test: fix new mine-loop tests, for parallel usage
The two new mine-loop tests for cancellation were depending on starting
with 0 jobs in the job-queue and no outside job-queue additions during
the test.  This is no longer a valid assumption when tests are run in
parallel.

The fix involves using a sleep instead of waiting for a specific number
of jobs.  It is less "clean" but it works for both sequential and
parallel test mode.